Densenet_121  implementation added

+++ run.sh
@@ -1,14 +1,38 @@
-## Cleanup
-pkill -9 tensorboard
-sleep 1
 
-rm -vrf output/*
+EXP_NAME='cifar10'
+CONFIG_FILE='configs/config_cifar10_small.json'
 
-tensorboard --logdir=output &
-sleep 1
+EXP_FILE='mains/main_densenet.py'
 
 
-## Run training
-# python mains/main_knifey_spoony.py -c configs/config_knifey_spoony.json
-# python mains/main_cifar10.py -c configs/config_cifar10.json
-python mains/main_densenet.py -c configs/config_cifar10.json
+# Train
+if [[ $1 -eq 1 ]]; then
+	echo "###########"
+	echo "#  Train  #"
+	echo "###########"
+	pkill -9 tensorboard
+	rm -vrf output/*
+	tensorboard --logdir=../output &
+	python $EXP_FILE -m 'train' -c $CONFIG_FILE
+
+# Val
+elif [[ $1 -eq 2 ]]; then
+	echo "################"
+	echo "#  Evaluation  #"
+	echo "################"
+	# pkill -9 tensorboard
+	# rm -rf "output"$EXP_NAME"/summary"
+	# tensorboard --logdir=../output &
+	python $EXP_FILE -m 'eval' -c $CONFIG_FILE
+
+# Test
+elif [[ $1 -eq 3 ]]; then
+	echo "################"
+	echo "#  Prediction  #"
+	echo "################"
+	python $EXP_FILE -m 'predict' -c $CONFIG_FILE
+
+else
+	echo "Unknown Argument"
+	echo "1 for Train, 2 for Evaluation and 3 for Prediction"
+fi
